Relax the assumptions about `lsb_release` output
The build.rs sources was contain a strong assumption about the
lsb_release output wrt the distribution version.

The output was expected to have Major.Minor form.

But for Debian (testing) the lsb_release command returns "testing" as
the version.

The previous assumption stops the build of the libsodium-sys crate.

Closes #32